+++ qb-engineer.core/Entities/PurchaseOrderLine.cs
@@ -5,14 +5,18 @@ public class PurchaseOrderLine : BaseEntity
     public int PurchaseOrderId { get; set; }
     public int PartId { get; set; }
     public string Description { get; set; } = string.Empty;
-    public int OrderedQuantity { get; set; }
-    public int ReceivedQuantity { get; set; }
+    // Phase 3 / WU-10 / F8-partial: quantities are decimal, not int. UoM-aware
+    // shops need fractional quantities — material-by-weight (lb, kg), by-time
+    // (hr), by-volume (gal, l). decimal(18, 4) = 4 fractional places, plenty
+    // for any reasonable UoM. Cases EDGE-DECIMAL-PRECISION-001 / -004.
+    public decimal OrderedQuantity { get; set; }
+    public decimal ReceivedQuantity { get; set; }
     public decimal UnitPrice { get; set; }
     public string? Notes { get; set; }
     public int? MrpPlannedOrderId { get; set; }
     public int? UomId { get; set; }
 
-    public int RemainingQuantity => OrderedQuantity - ReceivedQuantity;
+    public decimal RemainingQuantity => OrderedQuantity - ReceivedQuantity;
 
     public PurchaseOrder PurchaseOrder { get; set; } = null!;
     public Part Part { get; set; } = null!;
