+++ qb-engineer.core/Models/CreateSalesOrderRequestModel.cs
@@ -12,9 +12,11 @@ public record CreateSalesOrderRequestModel(
     decimal TaxRate,
     List<CreateSalesOrderLineModel> Lines);
 
+// Phase 3 / WU-10 / F8-partial — Quantity is decimal (was int). UoM-aware shops
+// need fractional quantities — material-by-weight, by-time, by-volume.
 public record CreateSalesOrderLineModel(
     int? PartId,
     string Description,
-    int Quantity,
+    decimal Quantity,
     decimal UnitPrice,
     string? Notes);
